Move test_amino_acid
